requirements.txt
ipython==7.16.1
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
Pillow==8.2.0
pytorch-lightning==1.1.8
PyYAML==5.3.1
scikit-learn==0.24.2
sklearn==0.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
